REFACTOR: clarify DSS utility module boundaries
